+++ src/test/java/lotto/TypeConverterTest.java
@@ -2,7 +2,7 @@ package lotto;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import lotto.converter.TypeConverter;
+import lotto.utils.TypeConverter;
 import lotto.message.ErrorMessage;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
